requirements.txt
numpy==1.15.1
Pillow==5.3.0
scipy==1.1.0
tensorflow-gpu==1.10.0
wget==3.2
seaborn==0.9.0
opencv-python
